+++ irs-api/src/main/java/org/eclipse/tractusx/irs/controllers/IrsController.java
@@ -45,15 +45,17 @@ import lombok.extern.slf4j.Slf4j;
 import org.eclipse.tractusx.irs.IrsApplication;
 import org.eclipse.tractusx.irs.component.Job;
 import org.eclipse.tractusx.irs.component.JobHandle;
-import org.eclipse.tractusx.irs.component.JobStatusResult;
 import org.eclipse.tractusx.irs.component.Jobs;
+import org.eclipse.tractusx.irs.component.PageResult;
 import org.eclipse.tractusx.irs.component.RegisterJob;
 import org.eclipse.tractusx.irs.component.enums.JobState;
 import org.eclipse.tractusx.irs.connector.job.IrsTimer;
 import org.eclipse.tractusx.irs.dtos.ErrorResponse;
 import org.eclipse.tractusx.irs.services.IrsItemGraphQueryService;
 import org.springdoc.api.annotations.ParameterObject;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -72,7 +74,8 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping(IrsApplication.API_PREFIX)
 @RequiredArgsConstructor
 @SuppressWarnings({ "PMD.AvoidDuplicateLiterals",
-                    "PMD.ExcessiveImports"
+                    "PMD.ExcessiveImports",
+                    "PMD.ShortVariable"
 })
 public class IrsController {
 
@@ -83,7 +86,7 @@ public class IrsController {
                security = @SecurityRequirement(name = "oAuth2", scopes = "profile email"),
                tags = { "Item Relationship Service" },
                description = "Register an IRS job to retrieve an item graph for given {globalAssetId}.")
-    @ApiResponses(value = { @ApiResponse(responseCode = "201", description = "Returns jobId of registered job.",
+    @ApiResponses(value = { @ApiResponse(responseCode = "201", description = "Returns id of registered job.",
                                          content = { @Content(mediaType = APPLICATION_JSON_VALUE,
                                                               schema = @Schema(implementation = JobHandle.class),
                                                               examples = { @ExampleObject(name = "complete",
@@ -112,24 +115,25 @@ public class IrsController {
     @IrsTimer("registerjob")
     @PostMapping("/jobs")
     @ResponseStatus(HttpStatus.CREATED)
+    @PreAuthorize("hasAuthority('view_irs')")
     public JobHandle registerJobForGlobalAssetId(final @Valid @RequestBody RegisterJob request) {
         return itemJobService.registerItemJob(request);
     }
 
-    @Operation(description = "Return job with optional item graph result for requested jobId.",
+    @Operation(description = "Return job with optional item graph result for requested id.",
                operationId = "getJobForJobId",
-               summary = "Return job with optional item graph result for requested jobId.",
+               summary = "Return job with optional item graph result for requested id.",
                security = @SecurityRequirement(name = "oAuth2", scopes = "profile email"),
                tags = { "Item Relationship Service" })
     @ApiResponses(value = { @ApiResponse(responseCode = "200",
-                                         description = "Return job with item graph for the requested jobId.",
+                                         description = "Return job with item graph for the requested id.",
                                          content = { @Content(mediaType = APPLICATION_JSON_VALUE,
                                                               schema = @Schema(implementation = Jobs.class),
                                                               examples = @ExampleObject(name = "complete",
                                                                                         ref = "#/components/examples/complete-job-result"))
                                          }),
                             @ApiResponse(responseCode = "206",
-                                         description = "Return job with current processed item graph for the requested jobId.",
+                                         description = "Return job with current processed item graph for the requested id.",
                                          content = { @Content(mediaType = APPLICATION_JSON_VALUE,
                                                               schema = @Schema(implementation = Jobs.class),
                                                               examples = @ExampleObject(name = "complete",
@@ -161,16 +165,17 @@ public class IrsController {
                                          }),
     })
     @IrsTimer("getjob")
-    @GetMapping("/jobs/{jobId}")
+    @GetMapping("/jobs/{id}")
+    @PreAuthorize("hasAuthority('view_irs')")
     public Jobs getJobById(
-            @Parameter(description = "JobId of the job.", schema = @Schema(implementation = UUID.class), name = "jobId",
+            @Parameter(description = "Id of the job.", schema = @Schema(implementation = UUID.class), name = "id",
                        example = "6c311d29-5753-46d4-b32c-19b918ea93b0") @Size(min = IrsAppConstants.JOB_ID_SIZE,
-                                                                               max = IrsAppConstants.JOB_ID_SIZE) @Valid @PathVariable final UUID jobId,
+                                                                               max = IrsAppConstants.JOB_ID_SIZE) @Valid @PathVariable final UUID id,
             @Parameter(
                     description = "<true> Return job with current processed item graph. <false> Return job with item graph if job is in state <COMPLETED>, otherwise job.") @Schema(
                     implementation = Boolean.class, defaultValue = "true") @RequestParam(value = "returnUncompletedJob",
                                                                                          required = false) final boolean returnUncompletedJob) {
-        return itemJobService.getJobForJobId(jobId, returnUncompletedJob);
+        return itemJobService.getJobForJobId(id, returnUncompletedJob);
     }
 
     @Operation(description = "Cancel job for requested jobId.", operationId = "cancelJobByJobId",
@@ -204,27 +209,28 @@ public class IrsController {
                                          }),
     })
     @IrsTimer("canceljob")
-    @PutMapping("/jobs/{jobId}")
+    @PutMapping("/jobs/{id}")
+    @PreAuthorize("hasAuthority('view_irs')")
     public Job cancelJobByJobId(
-            @Parameter(description = "JobId of the job.", schema = @Schema(implementation = UUID.class), name = "jobId",
+            @Parameter(description = "Id of the job.", schema = @Schema(implementation = UUID.class), name = "id",
                        example = "6c311d29-5753-46d4-b32c-19b918ea93b0") @Size(min = IrsAppConstants.JOB_ID_SIZE,
-                                                                               max = IrsAppConstants.JOB_ID_SIZE) @Valid @PathVariable final UUID jobId) {
+                                                                               max = IrsAppConstants.JOB_ID_SIZE) @Valid @PathVariable final UUID id) {
 
-        return this.itemJobService.cancelJobById(jobId);
+        return this.itemJobService.cancelJobById(id);
     }
 
-    @Operation(description = "Returns jobIds for requested job states.", operationId = "getJobIdsByJobStates",
-               summary = "Returns jobIds for requested job states.",
+    @Operation(description = "Returns paginated jobs with state and execution times.", operationId = "getJobsByJobStates",
+               summary = "Returns paginated jobs with state and execution times.",
                security = @SecurityRequirement(name = "oAuth2", scopes = "profile email"),
                tags = { "Item Relationship Service" })
     @ApiResponses(value = { @ApiResponse(responseCode = "200",
-                                         description = "List of job ids and status for requested job states.",
-                                         content = { @Content(mediaType = APPLICATION_JSON_VALUE, array = @ArraySchema(
-                                                 schema = @Schema(implementation = JobStatusResult.class), maxItems = Integer.MAX_VALUE),
+                                         description = "Paginated list of jobs with state and execution times for requested job states.",
+                                         content = { @Content(mediaType = APPLICATION_JSON_VALUE,
+                                                              schema = @Schema(implementation = PageResult.class),
                                                               examples = @ExampleObject(name = "complete",
                                                                                         ref = "#/components/examples/complete-job-list-processing-state"))
                                          }),
-                            @ApiResponse(responseCode = "400", description = "Returns jobIds for requested job states failed.",
+                            @ApiResponse(responseCode = "400", description = "Return jobs for requested job states failed.",
                                          content = { @Content(mediaType = APPLICATION_JSON_VALUE,
                                                               schema = @Schema(implementation = ErrorResponse.class),
                                                               examples = @ExampleObject(name = "error",
@@ -242,22 +248,16 @@ public class IrsController {
                                                               examples = @ExampleObject(name = "error",
                                                                                         ref = "#/components/examples/error-response-403"))
                                          }),
-                            @ApiResponse(responseCode = "404", description = "Job with the requested state not found.",
-                                         content = { @Content(mediaType = APPLICATION_JSON_VALUE,
-                                                              schema = @Schema(implementation = ErrorResponse.class),
-                                                              examples = @ExampleObject(name = "error",
-                                                                                        ref = "#/components/examples/error-response-404"))
-                                         }),
     })
     @IrsTimer("getjobbystate")
     @GetMapping("/jobs")
-    public List<JobStatusResult> getJobsByJobState(
+    @PreAuthorize("hasAuthority('view_irs')")
+    public PageResult getJobsByJobState(
             @Valid @ParameterObject @Parameter(description = "Requested job states.", in = QUERY,
-                                               explode = Explode.FALSE, array = @ArraySchema(
-                    schema = @Schema(implementation = JobState.class), maxItems = Integer.MAX_VALUE)) @RequestParam(value = "jobStates",
-                                                                                      required = false,
-                                                                                      defaultValue = "") final List<JobState> jobStates) {
-        return itemJobService.getJobsByJobState(jobStates);
+               explode = Explode.FALSE, array = @ArraySchema(schema = @Schema(implementation = JobState.class), maxItems = Integer.MAX_VALUE))
+            @RequestParam(value = "jobStates", required = false, defaultValue = "") final List<JobState> jobStates,
+            @ParameterObject final Pageable pageable) {
+        return itemJobService.getJobsByJobState(jobStates, pageable);
     }
 
 }